Add favorite icon
Added favorite icon by inserting favoriteViewModel into App bar according to class,
I'm not fond of this approach. when user taps on favorite icon we should emit an event,
and deal with this event outside of compose. In fact, we are on the Main screen still,
where we MainViewModel is responsible, in the refactoring, I'd be passing the db repo
in the constructor and handling adding and removing favorite there as well